# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BudgetWise API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountType = (typeof AccountType)[keyof typeof AccountType];

export const AccountType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
  investment: "investment",
  cash: "cash",
} as const;

export interface Account {
  id: number;
  name: string;
  type: AccountType;
  balance: number;
  currency: string;
  color?: string;
  createdAt: string;
}

export type CreateAccountBodyType =
  (typeof CreateAccountBodyType)[keyof typeof CreateAccountBodyType];

export const CreateAccountBodyType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
  investment: "investment",
  cash: "cash",
} as const;

export interface CreateAccountBody {
  name: string;
  type: CreateAccountBodyType;
  balance?: number;
  currency?: string;
  color?: string;
}

export type UpdateAccountBodyType =
  (typeof UpdateAccountBodyType)[keyof typeof UpdateAccountBodyType];

export const UpdateAccountBodyType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
  investment: "investment",
  cash: "cash",
} as const;

export interface UpdateAccountBody {
  name?: string;
  type?: UpdateAccountBodyType;
  balance?: number;
  currency?: string;
  color?: string;
}

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  income: "income",
  expense: "expense",
} as const;

export interface Category {
  id: number;
  name: string;
  icon?: string;
  color?: string;
  type: CategoryType;
}

export type CreateCategoryBodyType =
  (typeof CreateCategoryBodyType)[keyof typeof CreateCategoryBodyType];

export const CreateCategoryBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateCategoryBody {
  name: string;
  icon?: string;
  color?: string;
  type: CreateCategoryBodyType;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface Transaction {
  id: number;
  accountId: number;
  categoryId?: number;
  amount: number;
  type: TransactionType;
  description?: string;
  date: string;
  isRecurring?: boolean;
  accountName?: string;
  categoryName?: string;
  categoryColor?: string;
  categoryIcon?: string;
  createdAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface CreateTransactionBody {
  accountId: number;
  categoryId?: number;
  amount: number;
  type: CreateTransactionBodyType;
  description?: string;
  date: string;
  isRecurring?: boolean;
}

export type UpdateTransactionBodyType =
  (typeof UpdateTransactionBodyType)[keyof typeof UpdateTransactionBodyType];

export const UpdateTransactionBodyType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface UpdateTransactionBody {
  accountId?: number;
  categoryId?: number;
  amount?: number;
  type?: UpdateTransactionBodyType;
  description?: string;
  date?: string;
  isRecurring?: boolean;
}

export interface Budget {
  id: number;
  categoryId: number;
  categoryName?: string;
  categoryColor?: string;
  categoryIcon?: string;
  limitAmount: number;
  /** YYYY-MM */
  month: string;
  createdAt: string;
}

export interface CreateBudgetBody {
  categoryId: number;
  limitAmount: number;
  /** YYYY-MM format */
  month: string;
}

export interface UpdateBudgetBody {
  limitAmount?: number;
  month?: string;
}

export interface BudgetProgress {
  id: number;
  categoryId: number;
  categoryName?: string;
  categoryColor?: string;
  categoryIcon?: string;
  limitAmount: number;
  spentAmount: number;
  remainingAmount: number;
  percentUsed: number;
  month: string;
}

export interface Goal {
  id: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  targetDate?: string;
  icon?: string;
  color?: string;
  isCompleted: boolean;
  percentComplete: number;
  createdAt: string;
}

export interface CreateGoalBody {
  name: string;
  targetAmount: number;
  currentAmount?: number;
  targetDate?: string;
  icon?: string;
  color?: string;
}

export interface UpdateGoalBody {
  name?: string;
  targetAmount?: number;
  currentAmount?: number;
  targetDate?: string;
  icon?: string;
  color?: string;
  isCompleted?: boolean;
}

export type SubscriptionBillingCycle =
  (typeof SubscriptionBillingCycle)[keyof typeof SubscriptionBillingCycle];

export const SubscriptionBillingCycle = {
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
} as const;

export interface Subscription {
  id: number;
  name: string;
  amount: number;
  billingCycle: SubscriptionBillingCycle;
  nextBillingDate: string;
  categoryId?: number;
  categoryName?: string;
  isActive: boolean;
  color?: string;
  createdAt: string;
}

export type CreateSubscriptionBodyBillingCycle =
  (typeof CreateSubscriptionBodyBillingCycle)[keyof typeof CreateSubscriptionBodyBillingCycle];

export const CreateSubscriptionBodyBillingCycle = {
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
} as const;

export interface CreateSubscriptionBody {
  name: string;
  amount: number;
  billingCycle: CreateSubscriptionBodyBillingCycle;
  nextBillingDate: string;
  categoryId?: number;
  isActive?: boolean;
  color?: string;
}

export type UpdateSubscriptionBodyBillingCycle =
  (typeof UpdateSubscriptionBodyBillingCycle)[keyof typeof UpdateSubscriptionBodyBillingCycle];

export const UpdateSubscriptionBodyBillingCycle = {
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
} as const;

export interface UpdateSubscriptionBody {
  name?: string;
  amount?: number;
  billingCycle?: UpdateSubscriptionBodyBillingCycle;
  nextBillingDate?: string;
  categoryId?: number;
  isActive?: boolean;
  color?: string;
}

export interface DashboardSummary {
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  spendableAmount: number;
  savingsRate: number;
  accountCount: number;
  transactionCount: number;
  activeSubscriptionsCost: number;
}

export interface CategorySpending {
  categoryId: number;
  categoryName: string;
  categoryColor?: string;
  categoryIcon?: string;
  totalAmount: number;
  transactionCount: number;
  percentage: number;
}

export interface MonthlyFlow {
  /** YYYY-MM */
  month: string;
  income: number;
  expenses: number;
  net: number;
}

export type ListTransactionsParams = {
  accountId?: number;
  categoryId?: number;
  type?: ListTransactionsType;
  startDate?: string;
  endDate?: string;
  limit?: number;
  offset?: number;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export type ListTransactions200 = {
  data: Transaction[];
  total: number;
};

export type ListBudgetsParams = {
  /**
   * Month in YYYY-MM format
   */
  month?: string;
};

export type GetSpendingByCategoryParams = {
  /**
   * Month in YYYY-MM format, defaults to current month
   */
  month?: string;
};